fixing bug filter by category&tags

+++ app/category/controllers.js
@@ -4,15 +4,15 @@ const Category = require("./model");
 module.exports = {
   createCategory: async (req, res, next) => {
     try {
-      // // cek policy
-      // const policy = policyFor(req.user);
+      // cek policy
+      const policy = policyFor(req.user);
 
-      // if (!policy.can("create", "Category")) {
-      //   return res.json({
-      //     error: 1,
-      //     message: "You're not allowed to perform this action",
-      //   });
-      // }
+      if (!policy.can("create", "Category")) {
+        return res.json({
+          error: 1,
+          message: "You're not allowed to perform this action",
+        });
+      }
 
       const payload = req.body;
 
@@ -56,15 +56,15 @@ module.exports = {
   },
   updateCategory: async (req, res, next) => {
     try {
-      // // cek policy
-      // const policy = policyFor(req.user);
+      // cek policy
+      const policy = policyFor(req.user);
 
-      // if (!policy.can("update", "Category")) {
-      //   return res.json({
-      //     error: 1,
-      //     message: "You're not allowed to perform this action",
-      //   });
-      // }
+      if (!policy.can("update", "Category")) {
+        return res.json({
+          error: 1,
+          message: "You're not allowed to perform this action",
+        });
+      }
 
       const payload = req.body;
 
@@ -93,15 +93,15 @@ module.exports = {
   },
   deleteCategory: async (req, res, next) => {
     try {
-      // // cek policy
-      // const policy = policyFor(req.user);
+      // cek policy
+      const policy = policyFor(req.user);
 
-      // if (!policy.can("delete", "Category")) {
-      //   return res.json({
-      //     error: 1,
-      //     message: "You're not allowed to perform this action",
-      //   });
-      // }
+      if (!policy.can("delete", "Category")) {
+        return res.json({
+          error: 1,
+          message: "You're not allowed to perform this action",
+        });
+      }
 
       const category = await Category.findOneAndDelete({ _id: req.params.id });
