+++ app/tags/controllers.js
@@ -4,15 +4,15 @@ const Tag = require("./model");
 module.exports = {
   createTag: async (req, res, next) => {
     try {
-      // // cek policy
-      // const policy = policyFor(req.user);
+      // cek policy
+      const policy = policyFor(req.user);
 
-      // if (!policy.can("create", "Tag")) {
-      //   return res.json({
-      //     error: 1,
-      //     message: "You're not allowed to perform this action",
-      //   });
-      // }
+      if (!policy.can("create", "Tag")) {
+        return res.json({
+          error: 1,
+          message: "You're not allowed to perform this action",
+        });
+      }
 
       const payload = req.body;
 
@@ -53,15 +53,15 @@ module.exports = {
   },
   updateTag: async (req, res, next) => {
     try {
-      // // cek policy
-      // const policy = policyFor(req.user);
+      // cek policy
+      const policy = policyFor(req.user);
 
-      // if (!policy.can("update", "Tag")) {
-      //   return res.json({
-      //     error: 1,
-      //     message: "You're not allowed to perform this action",
-      //   });
-      // }
+      if (!policy.can("update", "Tag")) {
+        return res.json({
+          error: 1,
+          message: "You're not allowed to perform this action",
+        });
+      }
 
       const payload = req.body;
 
@@ -89,15 +89,15 @@ module.exports = {
   },
   deleteTag: async (req, res, next) => {
     try {
-      // // cek policy
-      // const policy = policyFor(req.user);
+      // cek policy
+      const policy = policyFor(req.user);
 
-      // if (!policy.can("delete", "Tag")) {
-      //   return res.json({
-      //     error: 1,
-      //     message: "You're not allowed to perform this action",
-      //   });
-      // }
+      if (!policy.can("delete", "Tag")) {
+        return res.json({
+          error: 1,
+          message: "You're not allowed to perform this action",
+        });
+      }
 
       const tag = await Tag.findOneAndDelete({ _id: req.params.id });
